+++ widget/include/widget/aspect_ratio_pixmap_label.h
